Bump MicroPython to 1.29.0 + Misc Fixes

+++ drivers/scd4x/i2c_hal.cpp
@@ -87,7 +87,7 @@ void sensirion_i2c_hal_free(void) {
  * @param count   number of bytes to read from I2C and store in the buffer
  * @returns 0 on success, error code otherwise
  */
-int8_t sensirion_i2c_hal_read(uint8_t address, uint8_t* data, uint16_t count) {
+int8_t sensirion_i2c_hal_read(uint8_t address, uint8_t* data, uint8_t count) {
     int result = __i2c->read_blocking(address, data, count, false);
     return result >= 0 ? 0 : result;
 }
@@ -104,7 +104,7 @@ int8_t sensirion_i2c_hal_read(uint8_t address, uint8_t* data, uint16_t count) {
  * @returns 0 on success, error code otherwise
  */
 int8_t sensirion_i2c_hal_write(uint8_t address, const uint8_t* data,
-                               uint16_t count) {
+                               uint8_t count) {
     int result = __i2c->write_blocking(address, data, count, false);
     return result >= 0 ? 0 : result;
 }
